+++ sentry-kotlin-multiplatform/src/commonAppleTest/kotlin/io/sentry/kotlin/multiplatform/BaseSentryScopeTest.kt
@@ -5,7 +5,6 @@ import cocoapods.Sentry.SentryScope as CocoaScope
 actual abstract class BaseSentryScopeTest {
     actual fun initializeScope(): Scope {
         val cocoaScope = CocoaScope()
-        val cocoaScopeProvider = CocoaScopeProvider(cocoaScope)
-        return Scope(cocoaScopeProvider)
+        return CocoaScopeProvider(cocoaScope)
     }
 }