+++ sentry-kotlin-multiplatform/src/commonJvmMain/kotlin/io/sentry/kotlin/multiplatform/SentryBridge.kt
@@ -64,8 +64,7 @@ internal actual object SentryBridge {
     private fun configureScopeCallback(scopeCallback: ScopeCallback): (JvmScope) -> Unit {
         return {
             val jvmScopeProvider = JvmScopeProvider(it)
-            val scope = Scope(jvmScopeProvider)
-            scopeCallback.invoke(scope)
+            scopeCallback.invoke(jvmScopeProvider)
         }
     }
 }