+++ sentry-kotlin-multiplatform/src/jvmMain/kotlin/io/sentry/kotlin/multiplatform/SentryInit.kt
@@ -3,11 +3,11 @@ package io.sentry.kotlin.multiplatform
 import io.sentry.Sentry
 import io.sentry.kotlin.multiplatform.extensions.toJvmSentryOptionsCallback
 
-// The context is unused here and only implemented to satisfy the expect
-public actual abstract class Context
-
 internal actual fun initSentry(context: Context?, configuration: OptionsConfiguration) {
     val options = SentryOptions()
     configuration.invoke(options)
     Sentry.init(options.toJvmSentryOptionsCallback())
 }
+
+// The context is unused here and only implemented to satisfy the expect
+public actual abstract class Context